requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0
plotly>=5.18.0
matplotlib>=3.8.0
openpyxl>=3.1.0
rapidfuzz>=3.5.0
pyyaml>=6.0.0
click>=8.1.0
